Optimazation file counter execution time

+++ FileSystemBrowser/FileSystemBrowser/Models/FilesInfo.cs
@@ -8,9 +8,9 @@ namespace FileSystemBrowser.Models
     public class FilesInfo
     {
         // File counters.
-        public int FileUnder10MbCounter { get; set; }
-        public int File10To50MbCounter { get; set; }
-        public int FileOver100MbCounter { get; set; }
+        public long FileUnder10MbCounter { get; set; }
+        public long File10To50MbCounter { get; set; }
+        public long FileOver100MbCounter { get; set; }
 
         // Files name container.
         public List<string> FileDirectory { get; set; }
